:bug:  chunkhash changed each time
