+++ .changeset/beige-mangos-admire.md
@@ -1,0 +1,5 @@
+---
+"whatplatformis": patch
+---
+
+chunkhash changed each time
